refactor: Rename import store module
