+++ src/store/index.js
@@ -2,8 +2,8 @@ import { combineReducers, createStore, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
 
 import authReducer from "./modules/auth";
-import lectureReducer from "./modules/lectures";
-import resourceReducer from "./modules/resources";
+import lectureReducer from "./modules/lecturesModule";
+import resourceReducer from "./modules/resourcesModule";
 import resourceCreateReducer from "./modules/resourceCreateModule";
 
 const rootReducer = combineReducers({
